requirements.txt
bidict==0.22.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
dnspython==2.4.1
docopt==0.6.2
eventlet==0.33.3
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.0.5
google-auth==2.22.0
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
passlib==1.7.4
pipreqs==0.4.13
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dotenv==1.0.0
python-engineio==4.5.1
python-socketio==5.8.0
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.6
yarg==0.1.9
